feat(runtime-c-api) Extract the `instance` module.
